database seeded

+++ src/Core/CharityHub.Core.Contract/Campaign/Queries/GetCampaignById/CampaignByIdResponseDto.cs
@@ -6,6 +6,6 @@ public class CampaignByIdResponseDto
     public int DonorCount { get; set; }
     public int RemainingDayCount { get; set; }
     public DateTime? StartDateTime { get; set; }
-    public int? ChargedAmountProgressPercentage { get; set; }
+    public decimal? ChargedAmountProgressPercentage { get; set; }
     public decimal? TotalAmount { get; set; }
 }
